Fix: Ensure task gets deleted on DELETE endpoint
Was testing and found out that it wasn't removing the task from the DB so debugged and found out it was not working properly.

Files:
- app/controllers/task_controller.py (modified)
- app/services/task_service.py (modified)

+++ app/controllers/task_controller.py
@@ -77,8 +77,7 @@ class TaskController:
         Return:
             (Response): A Flask response object containing a success message if the task was deleted, else an error message.
         """
-        task = TaskService.delete_task(task_id)
-
-        if not task:
-            abort(404, 'Task not found')
+        if TaskService.get_task(task_id) is None:
+            abort(404, "Task not found")
+        TaskService.delete_task(task_id)
         return {"message": "Task deleted successfully"}, 200
